pyproject.toml
[tool.poetry]
name = "pandas-tdd"
version = "0.1.0"
description = ""
authors = ["Grzegorz Kocjan <grzegorz@kocjan.me>"]
packages = [
    { include = "gradebook" }
]

[tool.poetry.dependencies]
python = "~3.10"
matplotlib = "^3.6.0"
numpy = "^1.23.3"
pandas = "^1.5.0"
scipy = "^1.9.1"
PyQt5 = "^5.15.7"

[tool.poetry.dev-dependencies]
Faker = "^14.2.0"
python-dateutil = "^2.8.2"
pesel = "^0.3.1"
pytest = "^7.1.3"
black = "^22.8.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
